add timeout and use optimized gzip, json encoder/decoder

+++ transfer_progress.go
@@ -8,5 +8,5 @@ type TransferProgress struct {
 	FileProcessed   int32
 	StartDate       string
 	startTime       *time.Time
-	ElapsedInSec    int32
+	ElapsedInSec    int64
 }
